@@ -1,53 +1,53 @@
 {
-    "name": "client-app",
-    "version": "2.5.4",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build",
-        "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "~8.2.14",
-        "@angular/common": "~8.2.14",
-        "@angular/compiler": "~8.2.14",
-        "@angular/core": "~8.2.14",
-        "@angular/forms": "~8.2.14",
-        "@angular/platform-browser": "~8.2.14",
-        "@angular/platform-browser-dynamic": "~8.2.14",
-        "@angular/router": "~8.2.14",
-        "@ng-bootstrap/ng-bootstrap": "6.2.0",
-        "bootstrap": "4.3.1",
-        "core-js": "^2.6.9",
-        "jquery": "3.4.1",
-        "popper": "^1.0.1",
-        "rxjs": "6.6.7",
-        "rxjs-compat": "6.3.3",
-        "tslib": "^1.10.0",
-        "zone.js": "~0.9.1"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "~0.803.29",
-        "@angular/cli": "~8.3.29",
-        "@angular/compiler-cli": "~8.2.14",
-        "@angular/language-service": "~8.2.14",
-        "@types/jasmine": "^2.8.16",
-        "@types/jasminewd2": "~2.0.3",
-        "@types/node": "~8.9.4",
-        "codelyzer": "^5.0.1",
-        "jasmine-core": "~2.99.1",
-        "jasmine-spec-reporter": "~4.2.1",
-        "karma": "~3.0.0",
-        "karma-chrome-launcher": "~2.2.0",
-        "karma-coverage-istanbul-reporter": "^2.0.5",
-        "karma-jasmine": "~1.1.2",
-        "karma-jasmine-html-reporter": "^0.2.2",
-        "protractor": "^5.4.2",
-        "ts-node": "~7.0.0",
-        "tslint": "~5.11.0",
-        "typescript": "~3.5.3"
-    }
+  "name": "client-app",
+  "version": "2.5.4",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "~9.1.13",
+    "@angular/common": "~9.1.13",
+    "@angular/compiler": "~9.1.13",
+    "@angular/core": "~9.1.13",
+    "@angular/forms": "~9.1.13",
+    "@angular/platform-browser": "~9.1.13",
+    "@angular/platform-browser-dynamic": "~9.1.13",
+    "@angular/router": "~9.1.13",
+    "@ng-bootstrap/ng-bootstrap": "6.2.0",
+    "bootstrap": "4.3.1",
+    "core-js": "^2.6.9",
+    "jquery": "3.4.1",
+    "popper": "^1.0.1",
+    "rxjs": "6.6.7",
+    "rxjs-compat": "6.3.3",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.901.15",
+    "@angular/cli": "~9.1.15",
+    "@angular/compiler-cli": "~9.1.13",
+    "@angular/language-service": "~9.1.13",
+    "@types/jasmine": "^2.8.16",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~3.0.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "^2.0.5",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "protractor": "^5.4.2",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~3.8.3"
+  }
 }
