@@ -5,7 +5,7 @@ import ptBr from '@angular/common/locales/pt';
 
 import { AppComponent } from './components/app/app.component';
 import { CommonModule, registerLocaleData, TitleCasePipe } from '@angular/common';
-import { HttpClient, HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
@@ -62,7 +62,6 @@ import { GameDetailTipEditComponent } from './components/game/game-detail-tip-ed
   imports: [
     BrowserModule,
     CommonModule,
-    HttpClient,
     HttpClientModule,
     FormsModule,
     NgbModule,
